+++ spec/checkers/postgresql/missing_unique_index_checker_spec.rb
@@ -131,5 +131,58 @@ RSpec.describe DatabaseConsistency::Checkers::MissingUniqueIndexChecker, :postgr
         )
       end
     end
+
+    context 'when index has different where scope' do
+      before do
+        define_database_with_entity do |table|
+          table.integer :account_id
+          table.boolean :is_default
+          table.boolean :deleted
+          table.index %i[account_id], unique: true, where: 'deleted = false'
+        end
+      end
+
+      specify do
+        expect(checker.report).to have_attributes(
+          checker_name: 'MissingUniqueIndexChecker',
+          table_or_model_name: klass.name,
+          column_or_attribute_name: 'account_id',
+          status: :fail,
+          error_message: nil,
+          error_slug: :missing_unique_index
+        )
+      end
+    end
+  end
+
+  context 'when uniqueness validation has different conditions scope' do
+    let(:attribute) { :account_id }
+    let(:klass) do
+      define_class do |klass|
+        klass.validates :account_id, uniqueness: { conditions: -> { where(deleted: false) } }
+      end
+    end
+
+    context 'when partial unique index with different where clause is provided' do
+      before do
+        define_database_with_entity do |table|
+          table.integer :account_id
+          table.boolean :is_default
+          table.boolean :deleted
+          table.index %i[account_id], unique: true, where: 'is_default = true'
+        end
+      end
+
+      specify do
+        expect(checker.report).to have_attributes(
+          checker_name: 'MissingUniqueIndexChecker',
+          table_or_model_name: klass.name,
+          column_or_attribute_name: 'account_id',
+          status: :fail,
+          error_message: nil,
+          error_slug: :missing_unique_index
+        )
+      end
+    end
   end
 end